requirements_audio.txt
librosa==0.9.2
matplotlib
numpy==1.24.4
numba
phonemizer
scipy
tensorboard
Unidecode
amfm_decompy
jieba
transformers
pypinyin
cn2an
gradio==3.50.2
av
mecab-python3
loguru
unidic-lite
cmudict
fugashi
num2words
PyYAML
requests
pyopenjtalk-prebuilt
jaconv
psutil
GPUtil
vector_quantize_pytorch
g2p_en
sentencepiece
pykakasi
langid
WeTextProcessing>=0.1.10; platform_system != "Windows"